+++ sdk/nodejs/cdn/getProfile.ts
@@ -23,11 +23,13 @@ import * as utilities from "../utilities";
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-azurerm/blob/master/website/docs/d/cdn_profile.html.markdown.
  */
-export function getProfile(args: GetProfileArgs, opts?: pulumi.InvokeOptions): Promise<GetProfileResult> {
-    return pulumi.runtime.invoke("azure:cdn/getProfile:getProfile", {
+export function getProfile(args: GetProfileArgs, opts?: pulumi.InvokeOptions): Promise<GetProfileResult> & GetProfileResult {
+    const promise: Promise<GetProfileResult> = pulumi.runtime.invoke("azure:cdn/getProfile:getProfile", {
         "name": args.name,
         "resourceGroupName": args.resourceGroupName,
     }, opts);
+
+    return pulumi.utils.liftProperties(promise, opts);
 }
 
 /**